requirements.txt
Django>=3.1
pandas
arrow
coreapi
Markdown
Pygments
djangorestframework
django-filter
django-guardian
drf-action-serializer
waitress
